+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and releases in Discovery project adheres to [Semantic Versioning](http://semver
 ### Added
 - Logic for Microform/Newspapers 'Request Me' link [#1217](https://github.com/ualbertalib/discovery/issues/1217)
 
+### Changed
+- More electronic access URLs being displayed [#877](https://github.com/ualbertalib/discovery/issues/877)
+- Refactored to not use javascript to display electronic access links [PR#1398](https://github.com/ualbertalib/discovery/pull/1398)
+
 ## [3.0.92] - 2018-11-16
 
 ### Fixed